refactor: introduce data dir
