+++ src/data/mod.rs
@@ -1,0 +1,3 @@
+pub mod artifacts;
+pub mod changelog;
+pub mod releases;
