+++ src/lib.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::uninlined_format_args)]
 
 pub mod config;
+pub mod data;
 pub mod errors;
 pub mod message;
 pub mod site;
